+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
-## [1.0.0] - 2020-07-12
+## [1.1.0] - 2020-07-12
+
+Added
+- Ability to run on non-production environments
+
+Fixed
+- Compatibility with `jekyll serve`
+
+## [1.0.0] - 2020-07-11
 
 Uploaded initial version